pyproject.toml
[tool.poetry]
name = "beauhurst-data"
version = "0.1.0"
description = ""
authors = ["konstantina54 <57228593+konstantina54@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "beauhurst_data"}]

[tool.poetry.dependencies]
python = "^3.10.5"
ruamel-yaml = "^0.17.21"
requests = "^2.28.1"
prefect = "^2.7.7"
prefect-blocks = {git = "git@github.com:technation-io/python-prefect-blocks.git"}


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
